feat(bpmn-moddle): make bpmn-moddle inherit from moddle
This commit makes bpmn-moddle an instance of moddle, exposing all
methods on moddle directly. In addition it requires the users to
instantiate the model per use (i.e. per bpmn diagram instance). This is
a prerequisite for ID matching.

bpmn-moddle adds the following methods on top of the moddle api:

* fromXML
* toXML

Internal changes:

* move to mocha/chai for testing

Related to bpmn-io/bpmn-js#92

BREAKING CHANGE:

* BpmnModdle must now be instantiated. Replace require('bpmn-moddle')
with require('bpmn-moddle')() to update
* BpmnModdle.instance() is gone in favor of inheritance. Use all moddle
methods directly on the bpmn-moddle instance.